Remove datastore interface and refactor singleton datastore class
